+++ package.json
@@ -67,7 +67,6 @@
     "url-loader": "^1.0.1",
     "husky": "^0.14.3",
     "jest": "^23.4.0",
-    "jest-serializer-path": "^0.1.15",
     "lint-staged": "^7.2.0",
     "memory-fs": "^0.4.1",
     "nsp": "^3.2.1",
@@ -90,11 +89,9 @@
     ]
   },
   "jest": {
+    "testURL": "http://localhost",
     "collectCoverageFrom": [
       "src/**/*.js"
-    ],
-    "snapshotSerializers": [
-      "jest-serializer-path"
     ]
   }
 }
